Terrain divisions (tiles) with a customizable size - Before was in correlation with the terrain size - Also, a try on draw large distances with meters as unit - Every dimension hardcoded at the moment

+++ source/terrain.h
@@ -11,9 +11,9 @@
 
 	struct TerrainOptions
 	{
-		size_t width;
-		size_t height;
-		size_t elevation;
+		int width;
+		int height;
+		int elevation;
 
 		char* heightmap_filename;
 		char* colormap_filename;
